requirements.txt
allure-behave==2.8.17
allure-python-commons==2.8.17
behave==1.2.6
cx-Oracle==8.0.0
lorem==0.1.1
nose==1.3.7
nose-allure-plugin==1.0.5
nose2==0.9.2
psycopg2==2.8.5
pytest==6.0.0
pytest-allure-adaptor==1.7.10
requests==2.24.0
selenium==3.141.0
webdriver-manager==3.2.1

